requirements.txt
# Core dependencies for simplified portfolio optimization system
# Single source of truth for all dependencies

# Data manipulation
pandas>=1.5.0
numpy>=1.21.0

# Financial data
yfinance>=0.2.0

# Portfolio optimization
cvxpy>=1.4.0

# Configuration
PyYAML>=6.0

# Machine Learning (simple implementations only)
scikit-learn>=1.0.0

# Web API (optional - for basic endpoints)
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0


# Visualization
matplotlib>=3.5.0
# Testing
pytest>=7.0.0
